+++ app/webpack.config.js
@@ -92,7 +92,7 @@ module.exports = (env = {}) => {
   const commonPlugins = [
     new webpack.DefinePlugin({
       __DEVELOPMENT__: JSON.stringify(!env.production),
-      __STATIC_BUILD__: JSON.stringify(env.staticBuild),
+      __STATIC_SITE__: JSON.stringify(env.staticSite),
       __REDUX_LOGGER_ENABLED__: devConfig.reduxLoggerEnabled,
       __REDUX_DEVTOOLS_ENABLED__: devConfig.reduxDevToolsEnabled,
       __LOG_LEVEL__: JSON.stringify(env.production ? 'warn' : devConfig.logLevel),
@@ -106,7 +106,7 @@ module.exports = (env = {}) => {
       template: 'src/client/index.hbs',
       templateParameters: {
         production: Boolean(env.production),
-        baseUrl: env.baseUrl,
+        baseUrl: process.env.BASE_URL || '',
       },
     }),
   ];
@@ -128,7 +128,9 @@ module.exports = (env = {}) => {
       ? productionEntry
       : developmentEntry,
     output: {
-      path: resolvePath('dist/static'),
+      path: env.staticSite
+        ? resolvePath('dist/static')
+        : resolvePath('dist/node/static'),
       pathinfo: !env.production,
       filename: 'bundle.js',
     },